SimpleHelix fix warnigns and a bug
Fixed Bug: in the constructor if SimpleHelix for the LCErrorMatrix the
uninitialised pointer was de-referenced to assign the values from the
constructor argument, this cannot work. Now we are just assigning the
object from the arguments to the member object.

Replace LCErrorMatrix pointer with object to make assignment work, also easier memory management
initialise members
unused function parameters